Support for get/set time info on a representation
Set time of a property in a RESQML2.2 approach.
Make default DataobjectRepo version to EML2.0 and RESQML2.0.1 in all cases
Replace some int to int64_t to avoid some potential range errors

+++ src/eml2/AbstractHdfProxy.h
@@ -145,7 +145,7 @@ namespace EML2_NS
 		 * @returns	The number of dimensions of the dataset if successful, otherwise returns a negative
 		 * 			value.
 		 */
-		DLL_IMPORT_OR_EXPORT unsigned int getDimensionCount(const std::string & datasetName) { return getElementCountPerDimension(datasetName).size(); }
+		DLL_IMPORT_OR_EXPORT uint64_t getDimensionCount(const std::string & datasetName) { return getElementCountPerDimension(datasetName).size(); }
 
 		/**
 		 * Get the number of elements in each dimension of an HDF5 dataset.
