+++ src/resqml2/DiscreteProperty.h
@@ -515,7 +515,7 @@ namespace RESQML2_NS
 		 *
 		 * @returns	True if there exists a minimum value, false if not.
 		 */
-		DLL_IMPORT_OR_EXPORT virtual bool hasMinimumValue(unsigned int index = 0) const = 0;
+		DLL_IMPORT_OR_EXPORT virtual bool hasMinimumValue(uint64_t index = 0) const = 0;
 
 		/**
 		 * @brief	Gets the minimum value of a non vector property or the minimum value of one given
@@ -532,7 +532,7 @@ namespace RESQML2_NS
 		 * 			than @p 0 for a non vector property or greater than the vector size for a vector
 		 * 			property).
 		 */
-		DLL_IMPORT_OR_EXPORT virtual int64_t getMinimumValue(unsigned int index = 0) const = 0;
+		DLL_IMPORT_OR_EXPORT virtual int64_t getMinimumValue(uint64_t index = 0) const = 0;
 
 		/**
 		 * @brief	Checks if a non vector property or a given value of a vector property has got a
@@ -544,7 +544,7 @@ namespace RESQML2_NS
 		 *
 		 * @returns	True if there exists a maximum value, false if not.
 		 */
-		DLL_IMPORT_OR_EXPORT virtual bool hasMaximumValue(unsigned int index = 0) const = 0;
+		DLL_IMPORT_OR_EXPORT virtual bool hasMaximumValue(uint64_t index = 0) const = 0;
 
 		/**
 		 * @brief	Gets the maximum value of a non vector property or the maximum value of one given
@@ -561,7 +561,7 @@ namespace RESQML2_NS
 		 * 			than @p 0 for a non vector property or greater than the vector size for a vector
 		 * 			property).
 		 */
-		DLL_IMPORT_OR_EXPORT virtual int64_t getMaximumValue(unsigned int index = 0) const = 0;
+		DLL_IMPORT_OR_EXPORT virtual int64_t getMaximumValue(uint64_t index = 0) const = 0;
 
 		/**
 		 * @brief	Sets the minimum value of a non vector property or the minimum value of one given
@@ -572,7 +572,7 @@ namespace RESQML2_NS
 		 * 					case) or zero-based index of the vector value for which we want to set the
 		 * 					minimum value (vector property case).
 		 */
-		DLL_IMPORT_OR_EXPORT virtual void setMinimumValue(int64_t value, unsigned int index = 0) const = 0;
+		DLL_IMPORT_OR_EXPORT virtual void setMinimumValue(int64_t value, uint64_t index = 0) const = 0;
 
 		/**
 		 * @brief	Sets the maximum value of a non vector property or the maximum value of one given
@@ -583,7 +583,7 @@ namespace RESQML2_NS
 		 * 					case) or zero-based index of the vector value for which we want to set the
 		 * 					maximum value (vector property case).
 		 */
-		DLL_IMPORT_OR_EXPORT virtual void setMaximumValue(int64_t value, unsigned int index = 0) const = 0;
+		DLL_IMPORT_OR_EXPORT virtual void setMaximumValue(int64_t value, uint64_t index = 0) const = 0;
 
 		/** The standard XML tag without XML namespace for serializing this data object. */
 		DLL_IMPORT_OR_EXPORT static const char* XML_TAG;
