+++ src/resqml2/Grid2dRepresentation.h
@@ -444,7 +444,7 @@ namespace RESQML2_NS
 		 *
 		 * @returns	The index of the origin on the supporting representation.
 		 */
-		DLL_IMPORT_OR_EXPORT virtual int getIndexOriginOnSupportingRepresentation() const = 0;
+		DLL_IMPORT_OR_EXPORT virtual uint64_t getIndexOriginOnSupportingRepresentation() const = 0;
 
 		/**
 		 * Gets the index of the origin of the current geometry on a particular dimension of the
@@ -473,7 +473,7 @@ namespace RESQML2_NS
 		 *
 		 * @returns	The number of nodes on the dimension @p dimension.
 		 */
-		DLL_IMPORT_OR_EXPORT virtual int getNodeCountOnSupportingRepresentation(unsigned int dimension) const = 0;
+		DLL_IMPORT_OR_EXPORT virtual uint64_t getNodeCountOnSupportingRepresentation(unsigned int dimension) const = 0;
 
 		/**
 		 * @brief	Gets the index offset of the nodes of the current geometry on a particular dimension
@@ -487,9 +487,9 @@ namespace RESQML2_NS
 		 *
 		 * @returns	The index offset on the dimension @p dimension.
 		 */
-		DLL_IMPORT_OR_EXPORT virtual int getIndexOffsetOnSupportingRepresentation(unsigned int dimension) const = 0;
+		DLL_IMPORT_OR_EXPORT virtual int64_t getIndexOffsetOnSupportingRepresentation(unsigned int dimension) const = 0;
 
-		DLL_IMPORT_OR_EXPORT unsigned int getPatchCount() const final {return 1;}
+		DLL_IMPORT_OR_EXPORT uint64_t getPatchCount() const final {return 1;}
 
 		/** The standard XML tag without XML namespace for serializing this data object. */
 		DLL_IMPORT_OR_EXPORT static const char* XML_TAG;
